# operator-terminal/gui/public/swagger/operator-terminal.yml
swagger: '2.0'
info:
  version: '1.0.0'
  title: Operator Terminal ptls-comm
  description: An API that allowes for administration of and PTT for the channels.
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
host: roip-mcu:3000
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /janusConfig:
    get:
      description: Returns the current Janus configuration
      produces:
        - application/json
      responses:
        '200':
          description: configuration response
          schema:
            $ref: '#/definitions/janusConfiguration'
        default:
          description: unexpected error
  /channels:
    get:
      description: Returns all available channels
      produces:
        - application/json
      responses:
        '200':
          description: channel response
          schema:
            type: array
            items:
              $ref: '#/definitions/channel'
        default:
          description: unexpected error
  /channels/{channelInternalName}:
    get:
      description: Returns the channel identified by the given internal name
      produces:
        - application/json
      parameters:
        - name: channelInternalName
          in: path
          description: Name of channel to fetch
          required: true
          type: string
      responses:
        '200':
          description: channel response
          schema:
            $ref: '#/definitions/channel'
        '400':
          description: parameter invalid
        '404':
          description: channel not found
        default:
          description: unexpected error
  /channels/{channelInternalName}/setChannelOccupied:
    put:
      description: Sets the channel to be occupied. This endpoint will be called by the radio interface.
      parameters:
        - name: channelInternalName
          in: path
          description: Name of channel to set to be occupied
          required: true
          type: string
        - name: occupiedBy
          in: query
          description: If available this field should contain the caller identification
          required: false
          type: string
      responses:
        '204':
          description: Successfully set
        '400':
          description: parameter invalid
        '404':
          description: channel not found
        default:
          description: unexpected error
  /channels/{channelInternalName}/setChannelReleased:
    put:
      description: Deactivates the radio device for a transmission. This endpoint will be called by the radio interface.
      parameters:
        - name: channelInternalName
          in: path
          description: Name of channel to set to be released
          required: true
          type: string
      responses:
        '204':
          description: Successfully set
        '400':
          description: parameter invalid
        '404':
          description: channel not found
        default:
          description: unexpected error
  /channels/{channelInternalName}/speechRequest:
    put:
      description: Turns in a speech request that will be validated and the radio's PTT will be activated, if positive.
      parameters:
        - name: channelInternalName
          in: path
          description: Internal name of channel
          required: true
          type: string
        - name: operatorTerminalId
          in: query
          description: Unique terminal ID
          required: true
          type: string
      responses:
        '204':
          description: now transmitting
        '400':
          description: parameter invalid
        '404':
          description: channel not found
        '503':
          description: channel temp. not available
        default:
          description: unexpected error
  /channels/{channelInternalName}/speechTerminated:
    put:
      description: Indicates that the radio call was terminated by the operator.
      parameters:
        - name: channelInternalName
          in: path
          description: Internal name of channel
          required: true
          type: string
        - name: operatorTerminalId
          in: query
          description: Unique terminal ID
          required: true
          type: string
      responses:
        '204':
          description: radio device informed
        '400':
          description: parameter invalid
        '404':
          description: channel not found
        default:
          description: unexpected error
definitions:
  janusConfiguration:
    type: object
    properties:
      janusHost:
        type: string
      janusPort:
        type: integer
        format: int32
      janusPath:
        type: string
  channel:
    type: object
    properties:
      channelInternalName:
        type: string
      name:
        type: string
      description:
        type: string
      operationMode:
        type: string
      triggerMode:
        type: string
      radioInterfaceWebHost:
        type: string
      radioInterfaceWebPort:
        type: integer
        format: int32